+++ src/main/java/com/github/zgraund/createreautomated/datagen/recipe/ModCompactingRecipeGen.java
@@ -2,6 +2,7 @@ package com.github.zgraund.createreautomated.datagen.recipe;
 
 import com.github.zgraund.createreautomated.CreateReAutomated;
 import com.github.zgraund.createreautomated.registry.ModItems;
+import com.simibubi.create.AllItems;
 import com.simibubi.create.api.data.recipe.CompactingRecipeGen;
 import com.simibubi.create.content.processing.recipe.HeatCondition;
 import net.minecraft.core.HolderLookup;
@@ -14,17 +15,18 @@ import javax.annotation.Nonnull;
 import java.util.concurrent.CompletableFuture;
 
 @SuppressWarnings("unused")
-public class ModCompactingRecipeGen extends CompactingRecipeGen {
+public final class ModCompactingRecipeGen extends CompactingRecipeGen {
     GeneratedRecipe DIAMOND = create(
             "bake_diamond",
             builder -> builder.require(ModItems.UNBAKED_DIAMOND).output(Items.DIAMOND).requiresHeat(HeatCondition.SUPERHEATED)
     );
     GeneratedRecipe
-            RAW_IRON = compactBits(ModItems.IRON_BIT, Items.RAW_IRON, HeatCondition.NONE),
             RAW_COPPER = compactBits(ModItems.COPPER_BIT, Items.RAW_COPPER, HeatCondition.NONE),
+            RAW_ZINC = compactBits(ModItems.ZINC_BIT, AllItems.RAW_ZINC, HeatCondition.NONE),
+            RAW_IRON = compactBits(ModItems.IRON_BIT, Items.RAW_IRON, HeatCondition.NONE),
             RAW_GOLD = compactBits(ModItems.GOLD_BIT, Items.RAW_GOLD, HeatCondition.HEATED);
 
-    public GeneratedRecipe compactBits(ItemLike bit, @Nonnull ItemLike output, HeatCondition heat) {
+    GeneratedRecipe compactBits(ItemLike bit, @Nonnull ItemLike output, HeatCondition heat) {
         return create(
                 BuiltInRegistries.ITEM.getKey(output.asItem()).getPath() + "_from_bits",
                 builder -> {