+++ src/main/java/com/github/zgraund/createreautomated/datagen/recipe/ModSequencedAssemblyRecipeGen.java
@@ -10,7 +10,7 @@ import net.minecraft.data.PackOutput;
 import java.util.concurrent.CompletableFuture;
 
 @SuppressWarnings("unused")
-public class ModSequencedAssemblyRecipeGen extends SequencedAssemblyRecipeGen {
+public final class ModSequencedAssemblyRecipeGen extends SequencedAssemblyRecipeGen {
     GeneratedRecipe DIAMOND = create(
             "unbaked_diamond_from_bits",
             builder ->