+++ changelog.md
@@ -8,6 +8,10 @@ Create ReAutomated 0.1.1
     - Added arm interaction point support
 - Added configuration to disable drill durability consumption
 
+#### Fixes
+
+- Added missing raw zinc recipe from bits
+
 --------------------------------------------------------
 Create ReAutomated 0.1.0
 --------------------------------------------------------